fix build issue
Move class method definition from header to source file,
because only inline methods are allowed to be in header files.
